OrTools.Sat.Model.Decode

Documentation

data Parser input a

A parsing function type synonym, to tidy up type signatures.

This synonym is used in three ways:

Applied to RawPrimitive, to parse primitive fields.

Applied to RawField, to parse fields which correspond to a single FieldNumber.

Applied to RawMessage, to parse entire messages.

Many of the combinators in this module are used to combine and convert between these three parser types.

Parsers can be combined using the Applicative, Monad and Alternative instances.

type RawPrimitive = ParsedField

Raw data corresponding to a single encoded key/value pair.

type RawMessage = IntMap RawField

Raw data corresponding to an entire message.

A Map from FieldNumbers to the those values associated with that FieldNumber.

parse :: Parser RawMessage a -> ByteString -> Either ParseError a

Parse a message (encoded in the raw wire format) using the specified Parser.

one :: Parser RawPrimitive a -> a -> Parser RawField a

This turns a primitive parser into a field parser by keeping the last received value, or return a default value if the field number is missing.

Used to ensure that we return the last value with the given field number in the message, in compliance with the protobuf standard.

The protocol buffers specification specifies default values for primitive types.

For example:

one float 0 :: Parser RawField Float

repeated :: Parser RawPrimitive a -> Parser RawField [a]

Parse a repeated field, or an unpacked collection of primitives.

Each value with the identified FieldNumber will be passed to the parser in the first argument, to be converted into a value of the correct type.

For example, to parse a packed collection of uint32 values:

repeated uint32 :: Parser RawField ([Word32])

or to parse a collection of embedded messages:

repeated . embedded' :: Parser RawMessage a -> Parser RawField ([a])

embedded' :: Parser RawMessage a -> Parser RawPrimitive a

Create a primitive parser for an embedded message from a message parser.

This parser does no merging of fields if multiple message fragments are sent separately.

at :: Parser RawField a -> FieldNumber -> Parser RawMessage a

Turn a field parser into a message parser, by specifying the FieldNumber.

This parser will fail if the specified FieldNumber is not present.

For example:

one float `at` fieldNumber 1 :: Parser RawMessage (Maybe Float)

int32 :: Parser RawPrimitive Int32

Parse a primitive with the int32 wire type.

double :: Parser RawPrimitive Double

Parse a double.

packedVarints :: Integral a => Parser RawPrimitive [a]

Parse a packed collection of variable-width integer values (any of int32, int64, sint32, sint64, uint32, uint64 or enumerations).